+++ packages/@expo/metro-config/src/transform-worker/__tests__/collect-dependencies-upstream.test.ts
@@ -13,10 +13,9 @@ import { transformFromAstSync } from '@babel/core';
 import generate from '@babel/generator';
 import * as babylon from '@babel/parser';
 import type { NodePath } from '@babel/traverse';
-import types from '@babel/types';
 import * as t from '@babel/types';
 import dedent from 'dedent';
-import nullthrows from 'nullthrows';
+import assert from 'node:assert';
 
 import type {
   Dependency,
@@ -42,9 +41,15 @@ const opts: Options = {
   unstable_allowRequireContext: false,
 };
 
+// asserts non-null
+function nullthrows<T extends object>(x: T | null, message?: string): NonNullable<T> {
+  assert(x != null, message);
+  return x;
+}
+
 const originalWarn = console.warn;
 
-beforeEach(() => {
+beforeAll(() => {
   console.warn = jest.fn();
 });
 
@@ -901,6 +906,9 @@ describe('import() prefetching', () => {
 });
 
 describe('Evaluating static arguments', () => {
+  beforeEach(() => {
+    jest.mocked(console.warn).mockReset();
+  });
   it('supports template literals as arguments', () => {
     const ast = astFromCode('require(`left-pad`)');
     const { dependencies, dependencyMapName } = collectDependencies(ast, opts);
@@ -1014,8 +1022,8 @@ describe('Evaluating static arguments', () => {
       `)
     );
   });
+
   it('warns at build-time when requiring non-strings with special option', () => {
-    jest.mocked(console.warn).mockReset();
     const ast = astFromCode('require(someVariable)');
     const opts: Options = {
       asyncRequireModulePath: 'asyncRequire',
@@ -1035,7 +1043,30 @@ describe('Evaluating static arguments', () => {
     );
     expect(console.warn).toHaveBeenCalledTimes(1);
     expect(console.warn).toHaveBeenCalledWith(
-      'Ambiguous import at line 1: require(someVariable). This module may not work as intended when deployed to a runtime.'
+      'Dynamic import at line 1: require(someVariable). This module may not work as intended when deployed to a runtime.'
+    );
+  });
+  it('warns at build-time when async import of non-strings with special option', () => {
+    const ast = astFromCode('import(someVariable)');
+    const opts: Options = {
+      asyncRequireModulePath: 'asyncRequire',
+      dynamicRequires: 'warn',
+      inlineableCalls: [],
+      keepRequireNames: true,
+      allowOptionalDependencies: false,
+      dependencyMapName: null,
+      unstable_allowRequireContext: false,
+    };
+    const { dependencies } = collectDependencies(ast, opts);
+    expect(dependencies).toEqual([]);
+    expect(codeFromAst(ast)).toEqual(
+      comparableCode(`
+        import(someVariable);
+      `)
+    );
+    expect(console.warn).toHaveBeenCalledTimes(1);
+    expect(console.warn).toHaveBeenCalledWith(
+      'Dynamic import at line 1: import(someVariable). This module may not work as intended when deployed to a runtime.'
     );
   });
 });
@@ -1482,14 +1513,14 @@ function adjustPosForCodeFrame(pos: { column: number; line: number }) {
   return pos ? { ...pos, column: pos.column + 1 } : pos;
 }
 
-function adjustLocForCodeFrame(loc: types.SourceLocation) {
+function adjustLocForCodeFrame(loc: t.SourceLocation) {
   return {
     start: adjustPosForCodeFrame(loc.start),
     end: adjustPosForCodeFrame(loc.end),
   };
 }
 
-function formatLoc(loc: types.SourceLocation, depIndex: number, dep: Dependency, code: any) {
+function formatLoc(loc: t.SourceLocation, depIndex: number, dep: Dependency, code: any) {
   return codeFrameColumns(code, adjustLocForCodeFrame(loc), {
     message: `dep #${depIndex} (${dep.name})`,
     linesAbove: 0,