+++ packages/babel-preset-expo/src/environment-restricted-imports.ts
@@ -14,16 +14,16 @@ export function environmentRestrictedImportsPlugin(
 ): babel.PluginObj {
   const { types: t } = api;
 
-  const isReactServer = api.caller(getIsReactServer) || api.caller(getIsServer);
+  const isAnyServerEnvironment = api.caller(getIsReactServer) || api.caller(getIsServer);
 
-  const forbiddenPackages = isReactServer
+  const forbiddenPackages = isAnyServerEnvironment
     ? FORBIDDEN_REACT_SERVER_IMPORTS
     : FORBIDDEN_CLIENT_IMPORTS;
 
   function checkSource(source: string, path: NodePath<any>) {
     forbiddenPackages.forEach((forbiddenImport) => {
       if (source === forbiddenImport) {
-        if (isReactServer) {
+        if (isAnyServerEnvironment) {
           throw path.buildCodeFrameError(
             `Importing '${forbiddenImport}' module is not allowed in a React server bundle. Add the "use client" directive to this file or one of the parent modules to allow importing this module.`
           );