+++ packages/expo-calendar/src/Calendar.ts
@@ -55,7 +55,6 @@ export type Calendar = {
   source: Source;
   /**
    * Type of calendar this object represents.
-   * Possible values: [`CalendarType`](#calendarcalendartype).
    * @platform ios
    */
   type?: CalendarType;
@@ -65,7 +64,6 @@ export type Calendar = {
   color: string;
   /**
    * Whether the calendar is used in the Calendar or Reminders OS app.
-   * Possible values: [`EntityTypes`](#calendarentitytypes).
    * @platform ios
    */
   entityType?: EntityTypes;
@@ -75,7 +73,6 @@ export type Calendar = {
   allowsModifications: boolean;
   /**
    * Availability types that this calendar supports.
-   * Possible values: Array of [`Availability`](#calendaravailability).
    */
   allowedAvailabilities: Availability[];
   /**
@@ -100,13 +97,11 @@ export type Calendar = {
   timeZone?: string;
   /**
    * Alarm methods that this calendar supports.
-   * Possible values: Array of [`AlarmMethod`](#calendaralarmmethod).
    * @platform android
    */
   allowedReminders?: AlarmMethod[];
   /**
    * Attendee types that this calendar supports.
-   * Possible values: Array of [`AttendeeType`](#calendarattendeetype).
    * @platform android
    */
   allowedAttendeeTypes?: AttendeeType[];
@@ -123,7 +118,6 @@ export type Calendar = {
   isSynced?: boolean;
   /**
    * Level of access that the user has for the calendar.
-   * Possible values: [`CalendarAccessLevel`](#calendarcalendaraccesslevel).
    * @platform android
    */
   accessLevel?: CalendarAccessLevel;
@@ -213,7 +207,7 @@ export type Event = {
   /**
    * Object representing rules for recurring or repeating events. Set to `null` for one-time events.
    */
-  recurrenceRule: RecurrenceRule;
+  recurrenceRule: RecurrenceRule | null;
   /**
    * Date object or string representing the time when the event starts.
    */
@@ -238,12 +232,10 @@ export type Event = {
   allDay: boolean;
   /**
    * The availability setting for the event.
-   * Possible values: [`Availability`](#calendaravailability).
    */
   availability: Availability;
   /**
    * Status of the event.
-   * Possible values: [`EventStatus`](#calendareventstatus).
    */
   status: EventStatus;
   /**
@@ -258,7 +250,6 @@ export type Event = {
   organizerEmail?: string;
   /**
    * User's access level for the event.
-   * Possible values: [`EventAccessLevel`](#calendareventaccesslevel).
    * @platform android
    */
   accessLevel?: EventAccessLevel;
@@ -337,9 +328,9 @@ export type Reminder = {
    */
   alarms?: Alarm[];
   /**
-   * Object representing rules for recurring or repeated reminders. Null for one-time tasks.
+   * Object representing rules for recurring or repeated reminders. `null` for one-time tasks.
    */
-  recurrenceRule?: RecurrenceRule;
+  recurrenceRule?: RecurrenceRule | null;
   /**
    * Date object or string representing the start date of the reminder task.
    */
@@ -380,17 +371,14 @@ export type Attendee = {
   name: string;
   /**
    * Role of the attendee at the event.
-   * Possible values: [`AttendeeRole`](#calendarattendeerole).
    */
   role: AttendeeRole;
   /**
    * Status of the attendee in relation to the event.
-   * Possible values: [`AttendeeStatus`](#calendarattendeestatus).
    */
   status: AttendeeStatus;
   /**
    * Type of the attendee.
-   * Possible values: [`AttendeeType`](#calendarattendeetype).
    */
   type: AttendeeType;
   /**
@@ -407,7 +395,7 @@ export type Attendee = {
 
 // @needsAudit
 /**
- * A method for having the OS automatically remind the user about an calendar item.
+ * A method for having the OS automatically remind the user about a calendar item.
  */
 export type Alarm = {
   /**
@@ -424,14 +412,13 @@ export type Alarm = {
   // @docsMissing
   structuredLocation?: AlarmLocation;
   /**
-   * Method of alerting the user that this alarm should use; on iOS this is always a notification.
-   * Possible values: [`AlarmMethod`](#calendaralarmmethod).
+   * Method of alerting the user that this alarm should use. On iOS this is always a notification.
    * @platform android
    */
   method?: AlarmMethod;
 };
 
-// @needsAudit @docsMissing
+// @docsMissing
 export type AlarmLocation = {
   /**
    * @platform ios
@@ -445,6 +432,10 @@ export type AlarmLocation = {
   };
 };
 
+// @docsMissing
+/**
+ * @platform ios
+ */
 export enum DayOfTheWeek {
   Sunday = 1,
   Monday = 2,
@@ -455,6 +446,10 @@ export enum DayOfTheWeek {
   Saturday = 7,
 }
 
+// @docsMissing
+/**
+ * @platform ios
+ */
 export enum MonthOfTheYear {
   January = 1,
   February = 2,
@@ -477,12 +472,11 @@ export enum MonthOfTheYear {
  * which is in turn based on [the iCal RFC](https://tools.ietf.org/html/rfc5545#section-3.8.5.3)
  * so you can refer to those to learn more about this potentially complex interface.
  *
- * Not all of the combinations make sense. For example, when frequency is `DAILY`, setting `daysOfTheMonth` makes no sense.
+ * Not all the combinations make sense. For example, when frequency is `DAILY`, setting `daysOfTheMonth` makes no sense.
  */
 export type RecurrenceRule = {
   /**
    * How often the calendar item should recur.
-   * Possible values: [`Frequency`](#calendarfrequency).
    */
   frequency: Frequency;
   /**
@@ -1227,11 +1221,16 @@ export const useRemindersPermissions = createPermissionHook({
   requestMethod: requestRemindersPermissionsAsync,
 });
 
+// @docsMissing
+/**
+ * platform ios
+ */
 export enum EntityTypes {
   EVENT = 'event',
   REMINDER = 'reminder',
 }
 
+// @docsMissing
 export enum Frequency {
   DAILY = 'daily',
   WEEKLY = 'weekly',
@@ -1239,14 +1238,24 @@ export enum Frequency {
   YEARLY = 'yearly',
 }
 
+// @docsMissing
 export enum Availability {
-  NOT_SUPPORTED = 'notSupported', // iOS
+  /**
+   * @platform ios
+   */
+  NOT_SUPPORTED = 'notSupported',
   BUSY = 'busy',
   FREE = 'free',
   TENTATIVE = 'tentative',
-  UNAVAILABLE = 'unavailable', // iOS
+  /**
+   * @platform ios
+   */
+  UNAVAILABLE = 'unavailable',
 }
 
+/**
+ * @platform ios
+ */
 export enum CalendarType {
   LOCAL = 'local',
   CALDAV = 'caldav',
@@ -1254,8 +1263,9 @@ export enum CalendarType {
   SUBSCRIBED = 'subscribed',
   BIRTHDAYS = 'birthdays',
   UNKNOWN = 'unknown',
-} // iOS
+}
 
+// @docsMissing
 export enum EventStatus {
   NONE = 'none',
   CONFIRMED = 'confirmed',
@@ -1263,6 +1273,10 @@ export enum EventStatus {
   CANCELED = 'canceled',
 }
 
+// @docsMissing
+/**
+ * @platform ios
+ */
 export enum SourceType {
   LOCAL = 'local',
   EXCHANGE = 'exchange',
@@ -1272,43 +1286,122 @@ export enum SourceType {
   BIRTHDAYS = 'birthdays',
 }
 
+// @docsMissing
 export enum AttendeeRole {
-  UNKNOWN = 'unknown', // iOS
-  REQUIRED = 'required', // iOS
-  OPTIONAL = 'optional', // iOS
-  CHAIR = 'chair', // iOS
-  NON_PARTICIPANT = 'nonParticipant', // iOS
-  ATTENDEE = 'attendee', // Android
-  ORGANIZER = 'organizer', // Android
-  PERFORMER = 'performer', // Android
-  SPEAKER = 'speaker', // Android
-  NONE = 'none', // Android
+  /**
+   * @platform ios
+   */
+  UNKNOWN = 'unknown',
+  /**
+   * @platform ios
+   */
+  REQUIRED = 'required',
+  /**
+   * @platform ios
+   */
+  OPTIONAL = 'optional',
+  /**
+   * @platform ios
+   */
+  CHAIR = 'chair',
+  /**
+   * @platform ios
+   */
+  NON_PARTICIPANT = 'nonParticipant',
+  /**
+   * @platform android
+   */
+  ATTENDEE = 'attendee',
+  /**
+   * @platform android
+   */
+  ORGANIZER = 'organizer',
+  /**
+   * @platform android
+   */
+  PERFORMER = 'performer',
+  /**
+   * @platform android
+   */
+  SPEAKER = 'speaker',
+  /**
+   * @platform android
+   */
+  NONE = 'none',
 }
 
+// @docsMissing
 export enum AttendeeStatus {
-  UNKNOWN = 'unknown', // iOS
-  PENDING = 'pending', // iOS
+  /**
+   * @platform ios
+   */
+  UNKNOWN = 'unknown',
+  /**
+   * @platform ios
+   */
+  PENDING = 'pending',
   ACCEPTED = 'accepted',
   DECLINED = 'declined',
   TENTATIVE = 'tentative',
-  DELEGATED = 'delegated', // iOS
-  COMPLETED = 'completed', // iOS
-  IN_PROCESS = 'inProcess', // iOS
-  INVITED = 'invited', // Android
-  NONE = 'none', // Android
+  /**
+   * @platform ios
+   */
+  DELEGATED = 'delegated',
+  /**
+   * @platform ios
+   */
+  COMPLETED = 'completed',
+  /**
+   * @platform ios
+   */
+  IN_PROCESS = 'inProcess',
+  /**
+   * @platform android
+   */
+  INVITED = 'invited',
+  /**
+   * @platform android
+   */
+  NONE = 'none',
 }
 
+// @docsMissing
 export enum AttendeeType {
-  UNKNOWN = 'unknown', // iOS
-  PERSON = 'person', // iOS
-  ROOM = 'room', // iOS
-  GROUP = 'group', // iOS
+  /**
+   * @platform ios
+   */
+  UNKNOWN = 'unknown',
+  /**
+   * @platform ios
+   */
+  PERSON = 'person',
+  /**
+   * @platform ios
+   */
+  ROOM = 'room',
+  /**
+   * @platform ios
+   */
+  GROUP = 'group',
   RESOURCE = 'resource',
-  OPTIONAL = 'optional', // Android
-  REQUIRED = 'required', // Android
-  NONE = 'none', // Android
+  /**
+   * @platform android
+   */
+  OPTIONAL = 'optional',
+  /**
+   * @platform android
+   */
+  REQUIRED = 'required',
+  /**
+   * @platform android
+   */
+  NONE = 'none',
 }
 
+// @docsMissing
+/**
+ * @platform android
+ */
 export enum AlarmMethod {
   ALARM = 'alarm',
   ALERT = 'alert',
@@ -1317,6 +1410,10 @@ export enum AlarmMethod {
   DEFAULT = 'default',
 }
 
+// @docsMissing
+/**
+ * @platform android
+ */
 export enum EventAccessLevel {
   CONFIDENTIAL = 'confidential',
   PRIVATE = 'private',
@@ -1324,6 +1421,10 @@ export enum EventAccessLevel {
   DEFAULT = 'default',
 }
 
+// @docsMissing
+/**
+ * @platform android
+ */
 export enum CalendarAccessLevel {
   CONTRIBUTOR = 'contributor',
   EDITOR = 'editor',
@@ -1336,6 +1437,10 @@ export enum CalendarAccessLevel {
   NONE = 'none',
 }
 
+// @docsMissing
+/**
+ * @platform ios
+ */
 export enum ReminderStatus {
   COMPLETED = 'completed',
   INCOMPLETE = 'incomplete',