+++ packages/expo-network/src/Network.types.ts
@@ -35,30 +35,40 @@ export enum NetworkStateType {
   /**
    * Active network connection over mobile data or [`DUN-specific`](https://developer.android.com/reference/android/net/ConnectivityManager#TYPE_MOBILE_DUN)
    * mobile connection when setting an upstream connection for tethering.
+   * @platform android
+   * @platform ios
    */
   CELLULAR = 'CELLULAR',
   /**
-   * Active network connection over WiFi.
+   * Active network connection over Wi-Fi.
+   * @platform android
+   * @platform ios
    */
   WIFI = 'WIFI',
   /**
    * Active network connection over Bluetooth.
+   * @platform android
    */
   BLUETOOTH = 'BLUETOOTH',
   /**
    * Active network connection over Ethernet.
+   * @platform android
+   * @platform ios
    */
   ETHERNET = 'ETHERNET',
   /**
-   * Active network connection over Wimax.
+   * Active network connection over WiMAX.
+   * @platform android
    */
   WIMAX = 'WIMAX',
   /**
    * Active network connection over VPN.
+   * @platform android
    */
   VPN = 'VPN',
   /**
    * Active network connection over other network connection types.
+   * @platform android
    */
   OTHER = 'OTHER',
 }
@@ -66,7 +76,7 @@ export enum NetworkStateType {
 // @needsAudit
 /**
  * Represents an event that provides the updated network state when there is a change in the network status.
- * This is passed as the argument to listeners registered with addNetworkStateListener.
+ * This is passed as the argument to listeners registered with [`addNetworkStateListener()`](#networkaddnetworkstatelistenerlistener).
  */
 export type NetworkStateEvent = NetworkState;